+++ packages/core/src/mixins/rowComp.js
@@ -45,7 +45,7 @@ import getStateClassnames from '../utils/getStateClassnames';
 import { statusPropTypes } from './withStatus';
 
 import Icon from '../Icon';
-import Text from '../Text';
+import Text, { VERTICAL_ORDER } from '../Text';
 
 import { STATUS_CODE } from '../StatusIcon';
 
@@ -59,6 +59,16 @@ const CENTER = 'center';
 const RIGHT = 'right';
 const REVERSE = 'reverse';
 export const ROW_COMP_ALIGN = { LEFT, CENTER, RIGHT, REVERSE };
+
+export const textPropTypes = {
+    align: PropTypes.string,
+    noGrow: PropTypes.bool,
+    verticalOrder: PropTypes.string,
+    basic: PropTypes.node,
+    aside: PropTypes.node,
+    tag: PropTypes.node,
+    bold: PropTypes.bool,
+};
 
 /**
  * Determine alignment for pre-configured <Text> based on
@@ -84,9 +94,8 @@ function determineTextAlign(compAlign, hasIcon) {
  * Get 'align' and 'noGrow' layout props for <Text>
  * as it would receive if rendered by rowComp().
  *
- * @param  {String} compAlign
- * @param  {Bool}   hasIcon
- * @return {Object} layoutProps
+ * @param  {string} compAlign
+ * @param  {boolean}   hasIcon
  */
 export function getTextLayoutProps(compAlign, hasIcon) {
     return {
@@ -98,6 +107,7 @@ export function getTextLayoutProps(compAlign, hasIcon) {
 const rowComp = ({
     defaultMinified = false,
     defaultAlign = LEFT,
+    defaultVerticalOrder = VERTICAL_ORDER.NORMAL,
 } = {}) => (WrappedComponent) => {
     const componentName = getComponentName(WrappedComponent);
 
@@ -108,12 +118,22 @@ const rowComp = ({
             minified: PropTypes.bool,
 
             // Text label props
-            align: PropTypes.oneOf(Object.values(ROW_COMP_ALIGN)),
+            align: PropTypes.oneOf([
+                ROW_COMP_ALIGN.LEFT,
+                ROW_COMP_ALIGN.CENTER,
+                ROW_COMP_ALIGN.RIGHT,
+                ROW_COMP_ALIGN.REVERSE,
+            ]),
+            verticalOrder: PropTypes.oneOf([
+                VERTICAL_ORDER.NORMAL,
+                VERTICAL_ORDER.REVERSE,
+            ]),
             icon: PropTypes.oneOfType([
                 PropTypes.string,
                 PropTypes.element
             ]),
             basic: PropTypes.node,
+            avatar: PropTypes.node,
             aside: PropTypes.node,
             tag: PropTypes.node,
             bold: PropTypes.bool,
@@ -133,8 +153,10 @@ const rowComp = ({
             minified: defaultMinified,
 
             align: defaultAlign,
+            verticalOrder: defaultVerticalOrder,
             icon: null,
             basic: null,
+            avatar: null,
             aside: null,
             tag: null,
             bold: false,
@@ -149,7 +171,7 @@ const rowComp = ({
         };
 
         static childContextTypes = {
-            align: PropTypes.oneOf(Object.values(ROW_COMP_ALIGN)),
+            textProps: PropTypes.shape(textPropTypes),
             ...statusPropTypes,
             // status,
             // statusOptions,
@@ -157,13 +179,38 @@ const rowComp = ({
         };
 
         getChildContext() {
-            const { align, status, statusOptions, errorMsg } = this.props;
+            const { status, statusOptions, errorMsg } = this.props;
+            const textProps = this.getTextProps();
 
             return {
                 status,
                 statusOptions,
                 errorMsg,
-                align, // for <TextInput>
+                // for <TextInput>
+                textProps,
+            };
+        }
+
+        getTextProps() {
+            const {
+                align,
+                verticalOrder,
+                icon,
+                basic,
+                aside,
+                tag,
+                bold,
+            } = this.props;
+
+            const textLayoutProps = getTextLayoutProps(align, !!icon);
+
+            return {
+                verticalOrder,
+                basic,
+                aside,
+                tag,
+                bold,
+                ...textLayoutProps,
             };
         }
 
@@ -180,34 +227,21 @@ const rowComp = ({
         }
 
         renderContent() {
-            const {
-                align,
-                icon,
-                basic,
-                aside,
-                tag,
-                bold,
-            } = this.props;
-            const textProps = { basic, aside, tag, bold };
-            const textLayoutProps = getTextLayoutProps(align, !!icon);
-
-            // Render icon element
             const iconElement = this.renderIconElement();
+            const textProps = this.getTextProps();
 
             return [
                 iconElement,
-                <Text
-                    key="comp-text"
-                    {...textProps}
-                    {...textLayoutProps} />
+                <Text key="comp-text" {...textProps} />
             ];
         }
 
         render() {
             const {
                 minified,
-
+                avatar,
                 align,
+                verticalOrder,
                 icon,
                 basic,
                 aside,
@@ -244,6 +278,7 @@ const rowComp = ({
 
             return (
                 <WrappedComponent className={wrapperClassName} {...otherProps}>
+                    {avatar}
                     {children || this.renderContent()}
                 </WrappedComponent>
             );