+++ packages/storybook/.storybook/webpack.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const autoprefixer = require('autoprefixer');
+const TerserPlugin = require('terser-webpack-plugin');
 
 const includePath = path.resolve(__dirname, '../..'); // gypcrete/packages
 const excludePath = /node_modules/;
@@ -9,6 +10,24 @@ module.exports = (baseConfig, env, defaultConfig) => {
     defaultConfig.resolve.modules.push(
         path.join(__dirname, '..')
     );
+
+    // By default storybook will minify js with default terser plugin
+    // https://webpack.js.org/configuration/optimization/#optimizationminimize
+    // That will mangle the function name so we couldn't read original
+    // component name in published storybook.
+    // Following config customize the TerserPlugin to remove the mangling.
+    // Note that this will make the total bundle size increase a little (2.9MB => 3.2MB).
+    // Consider storybook is for developer this should be fine.
+    if (defaultConfig.optimization && env === 'PRODUCTION') {
+        // eslint-disable-next-line no-param-reassign
+        defaultConfig.optimization.minimizer = [new TerserPlugin({
+            cache: true,
+            parallel: true,
+            terserOptions: {
+                mangle: false,
+            }
+        })];
+    }
 
     // Ref: Storybook webpack dev config https://git.io/fpJ6h
     const babelLoaderRule = defaultConfig.module.rules[0];